Added possibility to generate vertices in BUFFER
The buffer region xenon now has a Point
Sampler both for NEW and Next100. The
region can be selected in config as
BUFFER.

In addition, another region is defined
XENON. Which calls a Point Sampler
which combines the ACTIVE and BUFFER
regions to be able to generate
decays in the full xenon volume.
Useful for Xe-137 and Xe-135 studies.