Revert "Use overlay to change Geant4 version 10.6.2 -> 10.6.3"
This reverts commit c699bdadf8f3289814e79dd12c967f0592e0d119.

This was just a proof of concept to show how to install a version of a package
that is different from the one provided by `nixpkgs`. In the case of Geant4 (and
in the absence of our own binary cache) this results in a very lengthy
compilation time. So we switch back to the default G4 version, but keep the
proof of concept in history, for future reference.