Test cache use from users' perspective
That is, rather than using the pre-package GitHub action for activating the
cache, use the sort of commands would be used elsewhere.

Ideally humans would use home-manager to install `cachix`, but installing
home-manager in the action just for this purpose, would be too much hassle, so
we just use `nix-env -i`.