Add explanation comment
This comment should remind us that the actual value of the distance
between the gate and the sapphire windows is adjusted so that the
distance between the surface of the EP copper plate and the end of the
holders/teflon reflectors/poly is 24.8 mm, which is the value measured
by Sara in the step file. In other words, given ep_fc_distance_,
1432.1 mm is just adjusted to obtain that distance, in the current
version of the code.